input dropout
mostly useful to make model capable of scoring lower-order n-grams:
use dropout (p=0.95) during training, and pad the n-gram to score
with 0 embeddings (the embedding at null_index will be set to 0).

this is an alternative to obtaining the weighted average of all
input embeddings for padding, as done by (Vaswani et al. 2013)